requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
greenlet>=3.0.0
aiosqlite>=0.20.0
aiogram>=3.4.1
aiohttp-socks>=0.11.0
httpx>=0.27.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
python-multipart>=0.0.9
